test(conftest): moves common fixtures to conftest to reduce duplicate code
